Project/Src/requirements.txt
torch
torchvision
timm
facenet-pytorch
opencv-python-headless
pillow
numpy
tensorflow
gradio
